Fix reference to providers, use context instead
