+++ src/context/authentication/types.ts
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
 import { Result } from '../../common/Result';
-import {Update} from "../user/types";
+import { Update } from '../user/types';
 
 export type AuthenticationRepository = {
   login: (username: string, password: string) => Promise<Result<Error, string>>;
